chore(core): move utxo recovery tests to separate file
This allows cleaning them up further.

Issue: BG-34381
